requirements.txt
fastapi
uvicorn
psycopg2-binary
cx_Oracle
python-dotenv